feat: add python auto command for save
We can now check if the python file has a main block, so we can
execute the file and display the program execution after the file
be saved.

This is still a WIP now we need store buffers relative to python
files with main blocks and control those buffers visibility, and
the program execution output display as well.