refactor(lsp): use vim.lsp.config instead of lspconfig
This removes the depreciation warning that requiring the lspconfig
and using it directly for lsp configuration will be removed from
the project.

The vim.lsp.config requires also to call vim.lsp.enable of all
servers.

Now mason and mason-lspconfig are set as dependencies of
nvim-lspconfig and the real configuration is bootstraped into its
config function.